requirements.txt
numpy==1.24.3
scikit-learn==1.2.2
sentence-transformers==2.2.2
pinecone-client==2.2.4
python-dotenv==1.0.0
nltk==3.8.1
python-dateutil==2.8.2
fastapi==0.103.1
uvicorn==0.23.2
openai==1.3.0
jinja2==3.1.2
pydantic==2.4.2
starlette==0.27.0
email-validator==2.0.0
gradio==4.13.0
transformers==4.35.2
torch==2.1.0
accelerate==0.23.0
bitsandbytes==0.41.1
huggingface-hub==0.17.3

